part 2
for god's sake

+++ code/modules/mob/living/bot/medbot.dm
@@ -115,9 +115,8 @@
 		if(!isnull(reagent_glass))
 			to_chat(user, "<span class='notice'>There is already a container loaded.</span>")
 			return
-
-		user.drop_item()
-		O.loc = src
+		if(!user.drop(O, src))
+			return
 		reagent_glass = O
 		to_chat(user, "<span class='notice'>You insert [O].</span>")
 		return
@@ -339,7 +338,8 @@
 		switch(build_step)
 			if(0)
 				if(istype(W, /obj/item/device/healthanalyzer))
-					user.drop_item()
+					if(!user.drop(W))
+						return
 					qdel(W)
 					build_step++
 					to_chat(user, "<span class='notice'>You add the health sensor to [src].</span>")
@@ -348,7 +348,8 @@
 
 			if(1)
 				if(isprox(W))
-					user.drop_item()
+					if(!user.drop(W))
+						return
 					qdel(W)
 					to_chat(user, "<span class='notice'>You complete the Medibot! Beep boop.</span>")
 					var/turf/T = get_turf(src)
